requirements.txt
# Azure Functions Python Worker
# Pinned to specific versions for security and reproducibility
azure-functions==1.20.0
